Refactor: 필요없는 출력문 삭제 및 파일 정리
