+++ src/App.js
@@ -1,5 +1,6 @@
 import { Console, Random } from '@woowacourse/mission-utils';
-import { ERROR, MESSAGE } from './utils/constants.js';
+import { ERROR, MESSAGE } from './constants.js';
+import userInput from './UserInput.js';
 
 class App {
   computerNumber;
@@ -7,21 +8,16 @@ class App {
 
   constructor() {
     Console.print(MESSAGE.startMessage);
-    // this.computerNumber = this.randomGenerator();
   }
 
   async play() {
     this.computerNumber = this.randomGenerator();
     Console.print(this.computerNumber);
-    // this.userNumber = await this.getUserNumber();
-    // Console.print(this.userNumber);
-    // this.checkInputValidate();
     return this.start()
   };
 
   async start() {
     this.userNumber = await this.getUserNumber();
-    Console.print(this.userNumber);
     return this.checkInputValidate();
   }
 
@@ -50,13 +46,13 @@ class App {
   // 입력받은 수 유효성 검사
   checkInputValidate () {
     const IS_UNIQUE = (new Set(this.userNumber)).size;
+    const RE = new RegExp(/[1-9]{3}/g);
 
-    if (!this.userNumber.match(/[1-9]{3}/) || IS_UNIQUE !== 3) {
+    if (!RE.test(this.userNumber) || this.userNumber.length !== 3 || IS_UNIQUE !== 3) {
       throw new Error (ERROR.invalidInput);
     };
 
     this.userNumber =  this.userNumber.split('').map(Number);
-    // Console.print(this.userNumber)
     return this.countStrike();
   };
 
@@ -64,51 +60,39 @@ class App {
   countStrike() {
     const STRIKE = [...this.computerNumber].filter((x, idx) => this.userNumber[idx] === x).length;
     const BALL = [...this.computerNumber].filter(x => this.userNumber.includes(x)).length - STRIKE;
-    // Console.print(STRIKE, BALL,'llllll');
+    
     return this.strikeBall(STRIKE,BALL);
   };
 
   //스트라이크와 볼 개수 출력
   async strikeBall(strike, ball) {
-    // Console.print(strike, ball,'llllll');
     let answer = '';
     
     if (ball > 0) answer += `${ball}볼 `;
     if (strike > 0) answer += `${strike}스트라이크`;
     if (ball + strike === 0) answer = `낫싱`;
     Console.print(answer)
-
-    // if (strike === 3)  this.chooseRetry();
-    // return this.start();
-
-    // // if (strike === 3) this.chooseRetry();
-    // // else this.inputUserNumber();
+    
     if (strike !== 3) {
       return this.start();
     }
 
-    Console.print(MESSAGE.allStrike);
     const RETRY = await this.chooseRetry();
     return this.checkRetry(RETRY);
-    // this.checkRetry();
-
   };
 
   async chooseRetry() {
     try {
       Console.print(MESSAGE.allStrike);
       let retry = await Console.readLineAsync(MESSAGE.askRetry);
+
       return retry;
-      // this.retryOrExit(this.validateCheck.retryCheck(retry));
-      // retryOrExit(retry);
     } catch (error) {
       Console.print(error);
-    }
-
+    };
   };
 
   checkRetry(retry) {
-    Console.print(retry);
     if (retry !== "1" && retry !== "2") {
       throw new Error (ERROR.retryInput);
     };
@@ -123,11 +107,9 @@ class App {
       Console.print(MESSAGE.gameOver);
     };
   };
-}
+};
 
 const app = new App();
 app.play();
-
-
 
 export default App;
