+++ src/UserInput.js
@@ -1,0 +1,7 @@
+import { Console } from '@woowacourse/mission-utils';
+
+class UserInput {
+
+}
+
+export default UserInput;
